fixed docstring refs #2
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@26124 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/tools/traci/_person.py
@@ -224,7 +224,7 @@ class PersonDomain(Domain):
         self._connection._sendExact()
 
     def appendWaitingStage(self, personID, duration, description="waiting", stopID=""):
-        """appendWaitingStage(string, int, string, string)
+        """appendWaitingStage(string, float, string, string)
         Appends a waiting stage with duration in s to the plan of the given person
         """
         duration *= 1000
